Make the EventInfo class configurable (#65)
The `EventFactory` delegates to an event info class for publishing. The `EventInfo` class being used publishes messages with only a media type and URI.
Make the class used for building the event info a property of the factory, to give subclasses the possiblity to use a different event info class, with a different publishing behavior.